move constants to settings files and add optional chaining to environment variables
