+++ spa/src/components/authentication/logout.js
@@ -1,6 +1,6 @@
 import axios from 'ajax/axios';
 import { TOKEN } from 'ajax/endpoints';
-import { LS_USER, LS_CSRF_TOKEN } from 'constants/authConstants';
+import { LS_USER, LS_CSRF_TOKEN } from 'settings';
 
 async function logout() {
   try {
