+++ spa/src/components/donationPage/pageContent/DDonorAddress.js
@@ -5,7 +5,7 @@ import * as S from './DDonorAddress.styled';
 import { usePage } from 'components/donationPage/DonationPage';
 
 // Constants
-import { HUB_GOOGLE_MAPS_API_KEY } from 'constants/genericConstants';
+import { HUB_GOOGLE_MAPS_API_KEY } from 'settings';
 
 // Deps
 import { usePlacesWidget } from 'react-google-autocomplete';
