--- spa/src/constants/analyticsConstants.js
@@ -1,4 +1,0 @@
-export const HUB_ANALYTICS_APP_NAME = 'rev-engine-analytics';
-export const HUB_GA_V3_PLUGIN_NAME = 'ga-v3-hub';
-export const HUB_GA_V3_ID = window.ENV.HUB_V3_GOOGLE_ANALYTICS_ID;
-export const ORG_GA_V3_PLUGIN_NAME = 'ga-v3-org';
