--- spa/src/constants/authConstants.js
@@ -1,5 +1,0 @@
-export const LS_USER = 'REVENGINE_USER';
-export const LS_CONTRIBUTOR = 'REVENGINE_CONTRIBUTOR';
-export const LS_CSRF_TOKEN = 'CSRF_TOKEN';
-export const CSRF_HEADER = 'X-CSRFTOKEN';
-export const PASSWORD_RESET_URL = '/users/password-reset/';
