--- spa/src/constants/genericConstants.js
@@ -1,15 +1,0 @@
-// Google reCAPTCHA
-export const GRECAPTCHA_SCRIPT_URL = 'https://www.google.com/recaptcha/api.js';
-export const GRECAPTCHA_SITE_KEY = '6Lfuse8UAAAAAD9E6tCxKYrxO1IbnXp8IBa4u5Ri';
-export const HUB_GOOGLE_MAPS_API_KEY = window.ENV.HUB_GOOGLE_MAPS_API_KEY;
-
-// These subdomain labels will redirect to the org portal.
-export const ORG_PORTAL_SUBDOMAINS = ['', 'support'];
-
-export const HUB_STRIPE_API_PUB_KEY = window.ENV.HUB_STRIPE_API_PUB_KEY;
-export const REVENGINE_API_VERSION = window.ENV.REVENGINE_API_VERSION || 'v1';
-export const SALESFORCE_CAMPAIGN_ID_QUERYPARAM = window.ENV.SALESFORCE_CAMPAIGN_ID_QUERYPARAM || 'campaign';
-export const FREQUENCY_QUERYPARAM = window.ENV.FREQUENCY_QUERYPARAM || 'frequency';
-export const AMOUNT_QUERYPARAM = window.ENV.AMOUNT_QUERYPARAM || 'amount';
-export const CAPTURE_PAGE_SCREENSHOT = window.ENV.CAPTURE_PAGE_SCREENSHOT === 'true';
-export const STRIPE_API_VERSION = window.ENV.STRIPE_API_VERSION;
