+++ spa/src/utilities/isAuthenticated.js
@@ -1,4 +1,4 @@
-import { LS_CONTRIBUTOR, LS_USER } from 'constants/authConstants';
+import { LS_CONTRIBUTOR, LS_USER } from 'settings';
 
 function isAuthenticated(for_contributor) {
   if (for_contributor) return localStorage.getItem(LS_CONTRIBUTOR);
